Move airtable proxy tests to dedicated subfolder
