+++ api/tests/acceptance/application/airtable-proxy/airtable-proxy-controller-area_test.js
@@ -2,13 +2,13 @@ import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import nock from 'nock';
 import {
   airtableBuilder,
-  inputOutputDataBuilder,
   databaseBuilder,
   domainBuilder,
   generateAuthorizationHeader,
+  inputOutputDataBuilder,
   knex,
-} from '../../test-helper.js';
-import { createServer } from '../../../server.js';
+} from '../../../test-helper.js';
+import { createServer } from '../../../../server.js';
 
 describe('Acceptance | Controller | airtable-proxy-controller | create area translations', () => {
   beforeEach(() => {
