+++ api/tests/acceptance/application/airtable-proxy/airtable-proxy-controller-changelog_test.js
@@ -1,7 +1,7 @@
 import { describe, describe as context, expect, it } from 'vitest';
 import nock from 'nock';
-import { databaseBuilder, generateAuthorizationHeader } from '../../test-helper.js';
-import { createServer } from '../../../server.js';
+import { databaseBuilder, generateAuthorizationHeader } from '../../../test-helper.js';
+import { createServer } from '../../../../server.js';
 
 describe('Acceptance | Controller | airtable-proxy-controller-changelog', () => {
   describe('PATCH /api/airtable/changelog/Notes', () => {
