+++ api/tests/acceptance/application/airtable-proxy/airtable-proxy-controller-refresh-cache_test.js
@@ -4,12 +4,12 @@ import {
   airtableBuilder,
   databaseBuilder,
   domainBuilder,
-  inputOutputDataBuilder,
   generateAuthorizationHeader,
+  inputOutputDataBuilder,
   knex,
-} from '../../test-helper.js';
-import { createServer } from '../../../server.js';
-import * as config from '../../../lib/config.js';
+} from '../../../test-helper.js';
+import { createServer } from '../../../../server.js';
+import * as config from '../../../../lib/config.js';
 
 describe('Acceptance | Controller | airtable-proxy-controller-refresh-cache', () => {
